Implement basic support for a ``ConstantExpr`` to represent a floating point constant. The only way this is used right now is KModule constants that are of type ``llvm::ConstantFP`` are created using the ``APFloat`` constructor for ConstantExpr.
A ``ConstantExpr`` that represent a float prints itself differently.
Currently we support decimal and c99 hex float emission. When printing
expressions we currently emit c99 hex float because there is never
any precision loss. Unfortunately these are quite hard to read so I'm
not sure if this is the right design decision.