Implement FindArrayAckermannizationVisitor.
The purpose of this visitor is to find reads of arrays that can
ackermannized. The visitor doesn't actually make any modifications.
It just looks for opportunities to do so. The current implementation
is very conservative and will only say it can ackermannize simple
contiguous reads of non-constant arrays without updates or simple
array reads where the array has no updates. We can improve this
in the future.

We will use this to implement array ackermannization in the
Z3Solver/Z3Builder.

This not been tested an at all and is probably buggy...

The use of ``goto`` is a little unfortunately but I can't
think of an elegant and efficient way of handling failed matching
without using C++11 lambdas.