Add some plumbing to support multiple contiguous regions in an Array.
We don't support it right now because the ackermannization we are
looking for is mapping a single variable to a whole Array but in the
future it may be necessary to map multiple variables to contiguous
regions in an Array. When we do add support for this we will need
to be very careful how we handle ReadExpr because now we would need
to work out which variable we are mapping to.