Try to fix missing casts to Float in the Z3Builder.
I'm not sure if this is the right design choice. It would probably
be better if KLEE's Expr language was actually typed. Right now
with Floating point expression the semantics are that there is an
implicit conversion of a floating point expresion to the IEEE-754
bits when that expression gets used by operators that expect to work
with bitvectors.