Take the `Scalarize` pass from LLVM sources and use it to convert as much vectorized code as possible into scalar code so KLEE is able to a better job executing vectorized code.
Unfortunately the `Scalarize` pass was introduced in LLVM just
after the developers branched for LLVM 3.4 so I've had to copy the
sources into KLEE's source tree. I've also had to hack them a little
bit so it compiles but a glance it seems to be fixing the problems
I'm having with some benchmarks.

This isn't a complete solution because vectorized instructions
will still exist if vector types are passed across function call
boundaries. In that case we would really have to teach KLEE how
to handle this but for now this hack seems to be working.