Fix bug in handling ExtractElement instruction where we reading from the wrong index. This is a consequence of how KLEE evaluates constant vectors. It uses ConcatExpr and puts the 0th element in the most significant bits and adds the other elements to the right.
I'm not sure if this is really the write fix. We could change how
constants are evaluated and build the concat expression backwards
so elements are where we would expect them to be. However I'm worried
about other code that might depend on the current behaviour so its
easier to fix how ExtractElement is handled for now.