Implement dynamic solver timeout.
This is a complete hack but it seems to work. It can be enabled
by passing `-dynamic-solver-timeout` to KLEE. Passing
`-debug-only=dynamic_solver_timeout` will show debugging output
related to this feature.

The intuituion behind this idea is that sometimes we will terminate
a path due to solver timeout even though we have lots of global
time left to explore.

The current implementation currently takes the max of the remaining
executor time and the set solver timeout and recomputes this for
every solver call. In the future we might tweak this to take into
account things like the number of bugs found so far or the number
of paths created.