Implement native execution of casting x87 fp80 to unsigned and signed integers to work around incorrect behaviour of llvm::APFloat.
This is based off Aachen's ld-unnormal benchmark. Some seriously
weird stuff is happening here. There are some asserts that I've
had to hack because the computed value when run natively differs
depending on the whether Clang or gcc is used. Unfortunately this
means we won't be able to pass their test because we are compiling
KLEE with gcc which means we use gcc's behaviour for doing the
casting.