Fix calling `klee_make_symbolic()` when the pointer points to a `long double` and the size passed in is `sizeof(long double)`.
On x86_64 a `long double` is 80 bits in size but the alignment
128 bits. This caused the logic in the `klee_make_symbolic()`
handler to reject the call. Fix this by taking advantage of
the newly added `alignment` member on MemoryObjects and allow
the call to pass the check if

* The alignment is larger than the object size.
AND
* The requested size is equal to the alignment.

This requires that the pointer for MemoryObjects is correctly
aligned which was not done in the past.