Implement initial support in Z3Builder and Z3Solver for generating constraints over x87 fp80 Expr. We encode these as IEEE-754 binary floating point numbers with a 15-bit significand and 64-bit exponent. This gives us the right precision but because the binary encodings are different when creating a float from a bitvector we have to transform the encoding to the IEEE form from x87 fp80.
This requires that we generate side constraints that assert what the
value of the explicit significand integer bit has the correct value
that can be inferred from the other bits.

There are lots of other things that are missing but this is a very
promising start at supporting x87 fp80 constraints.