[REWORK of upstream patch] Fix bug where the incorrect amount of memory was allocated for padded types (e.g. `long double` on x86). Previously we would report an out of bound access if an attempt was made to access the padding.
This also fixes indexing into padded types which was incorrect
previously because padding was not included.

Two test cases are included. Unfortunately I couldn't
find a way to test `ConstantDataSequential` or `ConstantAggregateZero`
from a test case.

It also appears handling of vectors is broken in various ways
so those parts of the test cases are disabled until we can fix
them.