Misc ArrayAckermannization/Z3Builder fixes.
Fix bug where side constraints where not cleared before destroying
the Z3 context. The side constraints are stored in an
`std::vector<Z3ASTHandle>()` so the destructor on that was being
implicitly called which triggered calls to `Z3_dec_ref()` on a context
which had already been freed.

I'm suprised it took this long to be bitten by this.

--

Fix flaw in `FindArrayAckermannizationVisitor` where we would
fail to do ackermannization if we came across a structually
equivalent expression that was doing a read at a constant
offset we had already seen.