Teach the Z3Builder to convert expressions of floating point sort to bitvector sort avoiding use of `fp.to_ieee_bv`.
This function is an extension to the SMT-LIBv2.5 FloatingPoint theory.
For printing queries we may want to avoid this so this change is a step
in this direction.

This behaviour is controlled by a parameter to the `Z3Builder`
constructor. For actual solving the old behaviour is used but
`Z3Solver::getConstraintLog()` uses `Z3ConstraingLogConfig` to decide
what to do. The default behaviour is to use `fp.to_ieee_bv`.