Fix bug in alternative constraint encoding.
When avoiding to use of `fp.to_ieee_bv` with long doubles we generated
the wrong constraints. We should have returned the fresh 80-bit
bitvector but instead we returned that bitvector which certain bits
extracted.

Enabling this mode still breaks some tests because they rely on being
able to get the sign of a NaN.