Dynamic solver timeout version 2.
This is a complete hack but it seems to work. It can be enabled
by passing `-dynamic-solver-timeout` to KLEE. Passing
`-debug-only=dynamic_solver_timeout` will show debugging output
related to this feature.

The intuituion behind this idea is that sometimes we will terminate
a path due to solver timeout even though we have lots of global
time left to explore if we just used a simple static solver timeout.
We could also massively overshoot the executor's timeout if we get
a query near the timeout and use a large static solver timeout.

The idea is that during path exploration we compute

1. The remaining time allowed for path exploration
2. Compute the max of that and `DynamicSolverTimeoutMinQueryTimeDuringPathExploration`

We then use that as the solver timeout. By default
`DynamicSolverTimeoutMinQueryTimeDuringPathExploration` is 0.0 so the
max operation is a no-op.

`DynamicSolverTimeoutMinQueryTimeDuringPathExploration` can be set on
the command line with
`-dynamic-solver-timeout-min-query-time-during-path-exploration`.

The purpose of `DynamicSolverTimeoutMinQueryTimeDuringPathExploration`
is to allow us to have a lower bound on query timeout during path
exploration if we need it. In the very original of the dynamic
solver timeout the core solver timeout was used instead but
there is no good reason to conflate these two values because
they are different.

During test case generation if `DynamicSolverTimeoutTestCaseGenMaxTime`
is less than or equal to 0 we just allow an unlimited amount of time
for solver queries during test case generation.

`DynamicSolverTimeoutTestCaseGenMaxTime` can be set on the command line
with `-dynamic-solver-timeout-max-test-gen-time`.

If `DynamicSolverTimeoutTestCaseGenMaxTime` is positive then it
is considered to be the total amount of time allowed to do test
case generation and so the time to allow of each query is computed
by

1. Dividing this time equally amoung the remaining states.
2. Taking the max of that value with
`DynamicSolverTimeoutMinQueryTimePerTestCase`.

We then use that as the solver timeout. By default
`DynamicSolverTimeoutMinQueryTimePerTestCase` is 0.0 so
the max() is a no-op.

The `DynamicSolverTimeoutMinQueryTimePerTestCase` parameter exists to
allow us to have a lower bound on the query timeout during test case
generation.

`DynamicSolverTimeoutMinQueryTimePerTestCase` can be set on the command
line with `-dynamic-solver-timeout-min-query-time-per-test-case`.