Fix really stupid bug in dynamic solver.
Because we do max(v, 0.0) where `v` is some value we may set the
solver timeout to 0.0. The semantics of this in KLEE actually mean
run with **NO TIMEOUT** which is not what we want. What we actually
want is for the solver to basically to fail very quickly.

This now "sort of" fixed by actually just avoiding solver invocation
completely when this scenario occurs.